requirements.txt
pandas==1.0.1
Pillow==7.2.0
TwitterAPI==2.5.12
